Move pcie/windows to alveo sub-folder
