+++ src/runtime_src/core/pcie/windows/alveo/system_windows.h
@@ -1,21 +1,9 @@
-/**
- * Copyright (C) 2019 Xilinx, Inc
- *
- * Licensed under the Apache License, Version 2.0 (the "License"). You may
- * not use this file except in compliance with the License. A copy of the
- * License is located at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ * Copyright (C) 2019-2021 Xilinx, Inc. All rights reserved.
  */
-
-#ifndef SYSTEM_WINDOWS_H
-#define SYSTEM_WINDOWS_H
+#ifndef SYSTEM_WINDOWS_ALVEO_H
+#define SYSTEM_WINDOWS_ALVEO_H
 
 #include "pcie/common/system_pcie.h"
